shill: use Device::link_name() where interface name is expected
Device::link_name() returns the interface name associated with the
Device object. Device::UniqueName() returns a unique name for the Device
object, which happens to use the interface name in the current
implementation but this is no guarantee on that. This CL updates the
code to use Device::link_name() where the interface name associated with
the Device object is expected.

BUG=None
TEST=Run unit tests.

Change-Id: I25f096f4117583f9777bac18db98f45e24dba167
Reviewed-on: https://chromium-review.googlesource.com/715176
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>