chromeos-config: Fixing broken unit tests in cros/master
The commit crrev.com/c/704267 added a firmware node to caroline in the
test.dts file. The C++ library would open the firmware node, get the
property 'bcs-overlay' (which returned nullptr). That nullptr was casted
to a const char* and passed to a std::string constructor, resulting in a
segfault and dump.

This is a partial fix. It does not address the overarching issue: why
wasn't the unit test run for the host during CQ (see linked bug).

BUG=chromium:773341
TEST=FEATURES=test sudo -E emerge chromeos-base/chromeos-config-tools

Change-Id: Ifcd3083de39fec73dcb1a287597a7fa66160b6c2
Reviewed-on: https://chromium-review.googlesource.com/713763
Commit-Ready: Alec Thilenius <athilenius@chromium.org>
Tested-by: Alec Thilenius <athilenius@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>