login: Consolidate common storage/retrieval code
Points all Store* and Retrieve* methods to the corresponding *Ex
methods. Gets rid of a lot of very similar error handling code.
The behavior should be the same, the only thing that changes is the
returned error code that changed from
  kSessionNotStarted (user sessions) and
  kInvalidAccount (device local accounts)
to a common
  kGetServiceFail.

BUG=chromium:765644
TEST=emerge-amd64-generic chromeos-login
      cros_run_unit_tests --board=amd64-generic --packages chromeos-login

Change-Id: I5e982d6a181e47a602fb72765a98e824ffbc72d0
Reviewed-on: https://chromium-review.googlesource.com/719099
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>