chromeos-config: Use our own phandle map in CrosConfig
It is odd that looking up a phandle within CrosConfig returns an fdt.Node
instead of a CrosConfig.Node. This cannot work at present since
CrosConfig.Node objects are not  created for all nodes.

Fix this by building our own phandle dict at the start.

Note: We should consider renaming CrosConfig.Node as it is a bit confusing
that they are called the same.

BUG=chromium:771187
TEST=FEATURES=test sudo -E emerge chromeos-base/chromeos-config-tools
Change-Id: I30341116d5119435c94e917841e4c76a4471a23c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/720442
Reviewed-by: Jason Clinton <jclinton@chromium.org>