chromeos-config: Move GetElement out of the schema impl
This function doesn't really belong here. We have tried to get around it
by avoiding accessing any nodes directly, but it does check the node path
in one case.

It seems better to put it with the config validator, which is the piece
which should understand the node/property structure. This will allow us to
do more checks on this without the validator reaching into places it should
not.

BUG=chromium:761265
TEST=FEATURES=test emerge-coral chromeos-config-tools

Change-Id: Ifefc3e665943a3435e6c8a31b4c89f824eb364ec
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/721014
Reviewed-by: Jason Clinton <jclinton@chromium.org>