chromeos-config: Move Present() into ConfigValidator
This function works OK in the schema file since it has a fairly simple job
and we can pass in the required list of siblings. However we plan to add
support for making presense of one element dependent on present of another
element which is not necessarily in the same node. This involves looking
around the Node/Property structure, so it seems best to move this function
into the validator.

BUG=chromium:761265
TEST=FEATURES=test emerge-coral chromeos-config-tools

Change-Id: I9c82f9524bd94ae83eb1ee0befc4e44cfd8e5443
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/721016
Reviewed-by: Jason Clinton <jclinton@chromium.org>