Revert "chromeos-config: Added get_firmware_uris flag to cros_config_host"
We have moved cros_config_host to Python, so we don't need this tool now.
A follow-up CL will remove it once we have dropped it from eclasses.

This is a partial revert, leaving the test.dts changes which are used in
the cros_config_host_py tests.

This reverts commit 01cc9eb4e0332d72d83a98b9b2d2d582e9895951.

Change-Id: Ib37316beb9ebc2b1b8ff585f36e0af95ac845990
BUG=chromium:761265
TEST=FEATURES=test emerge-coral chromeos-config-tools
Need to run trybots also
Reviewed-on: https://chromium-review.googlesource.com/721018
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>