vm_tools: maitred: Implement proper shut down
Implement proper VM shut down by having maitred first send a SIGTERM to
all processes, wait up to 5 seconds for them to exit, and kill any
stragglers with SIGKILL.  maitred will then shut down the gRPC server
before issuing a reboot() syscall.

In practice, it takes ~1 second from the point that the server receives
the request to the point that the reboot syscall is issued.  In the
worst case this can take an infinite amount of time if the gRPC server
is unable to shut down.

BUG=none
TEST=manual.  Start some long running processes and initiate a shutdown.
     See that the RPC is successful and the VM shuts down properly with
     no warning or error logs.

Change-Id: I4df036b16035fa30561fc3210c622adb402d37e7
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/725520