chromeos-config: Show a simple file tree
This is just a starting point showing how files in the config can be
listed in a user-friendly form. It allows you to see the size of each file
and also those which are missing. Future work could involve actually
copying the files to a directory so that can be examined/compared.

BUG=chromium:774322
TEST=PYTHONPATH=~/cosarm python cros_config_host_py/cros_config_host_py \
	~/cosarm/chroot/build/reef-uni/usr/share/chromeos-config/config.dtb \
	file-tree ~/cosarm/chroot/build/reef-uni
Size      Path
          /
             etc/
                cras/
                   basking/
1992                  bxtda7219max
2500                  dsp.ini
                   electro/
2027                  bxtda7219max
1728                  dsp.ini
...

Change-Id: I4754a3b2ac78860dd6ee590061f9f5e62a622e74
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/726162
Reviewed-by: Jason Clinton <jclinton@chromium.org>