cryptohome: don't sign or create boot-lockbox keys if finalized
Boot lockbox doesn't allow signing after the lockbox is finalized
(the key won't load by design). Do a check at the beginning of
signing to avoid performing further operations in case it is known
to be finalized.

Similarly, if a boot-lockbox key doesn't exist yet, it should be
created only when the lockbox is not finalized yet. To avoid
potential races, add a check after creating the key but before
saving or using it, and abort if the lockbox is finalized at this
point.

BUG=b:67907568
TEST=unit tests (new added)

Change-Id: I4ddd8a9d48599b9738de377e7d2a303bcab074b9
Reviewed-on: https://chromium-review.googlesource.com/726320
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Xiaoyong Zhou <xzhou@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Xiaoyong Zhou <xzhou@chromium.org>