cros-disks: remove IsAlive D-Bus method
The IsAlive D-Bus method is provided to allow a D-Bus client of
cros-disks to check if the cros-disks D-Bus service is alive, which is
only used in the platform_CrosDisksDBus test. Given that a D-Bus client
can use other D-Bus methods exposed by cros-disks to determine if
cros-disks is alive, this CL simplfy remove the IsAlive D-Bus method.

BUG=chromium:767645
CQ-DEPEND=CL:726750
TEST=Run platform_CrosDisks{DBus,Filesystem,Format,Rename} tests.

Change-Id: Iec77990c0d270dea76f523674d6801eb9111c359
Reviewed-on: https://chromium-review.googlesource.com/727120
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>