intel: ipu3: Add error handling when mClient is nullptr
1. When the algo is not connected,
the Intel3AClient::getInstance() will be nullptr.
So we need to add nullptr checking for mClient
before using it.
2. If the camera service gets 0 camera number,
the hal lib will be reloaded.
So calling Intel3AClient::getInstance()
in hal_get_number_of_cameras() checks
if the connecting to the remote 3a is successful.

BUG=b:67882568
TEST=Build and deploy package to poppy and camera works

Change-Id: Id265c71cff1c4f19623c186e97c44bf8a578f487
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/729226
Commit-Ready: WeiX Meng <weix.meng@intel.com>
Tested-by: WeiX Meng <weix.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Liang L Yang <liang.l.yang@intel.com>