validate_config: Support validation of multiple fragments
It is useful to split the config into multiple .dtsi files to allow
different teams to work on them without interfering with others. We still
want to validate all the config as a whole, but need to show the filename
where the error occurred, even it is in a fragment.

Add -p option which accepts a list of .dtsi files to validate. It uses the
/include/ directive to compile them, thus ensuring error line numbers are
useful. Then it runs the resulting .dtb through the validator.

BUG=chromium:761265
TEST=PYTHONPATH=~/cosarm python validate_config_unittest.py

Change-Id: Ie0edd040da1bcec556923ab4c1c89d0f0ef00588
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/728204