chromeos-config: Access the schema from libcros_config_host
We will use the schema to access information about properties, in
particular the upcoming target_dir property. Set this up.

For running locally (within the platform2/chromeos-config directory) we
need to fiddle the path to permit access to the validator. Once everything
is in one place we will be able to drop this. See crbug.com/774492

BUG=chromium:769575
TEST=PYTHONPATH=~/cosarm \
   python cros_config_host_py/libcros_config_host_unittest.py;
PYTHONPATH=~/cosarm \
   python validate/validate_config_unittest.py

Change-Id: I0739facf88bd710ad0c19c69dc8c8fb9424e2306
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/729136
Reviewed-by: Jason Clinton <jclinton@chromium.org>