chromeos-config: Add a target directory to the schema
Some files need to be installed in the root. Rather than having the path
for that in two places (the ebuild and the init script, for example), we
can build this information into cros_config. Then cros_config_host and
cros_config can return the absolute path, both from the same source.

Add a new schema property for this.

BUG=chromium:769575
TEST=PYTHONPATH=~/cosarm \
   python cros_config_host_py/libcros_config_host_unittest.py;
PYTHONPATH=~/cosarm \
   python validate/validate_config_unittest.py

Change-Id: I5fcb072109c5515e5ed64273abb81ca2a99da260
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/729137
Reviewed-by: Jason Clinton <jclinton@chromium.org>