trunks, tpm_manager: prepare for ownership earlier
Taking TPM ownership during OOBE can only be started when
Chrome sends an appropriate command to cryptohomed. However,
there are several operations that can be done on the TPM,
and which are needed for taking ownership, before that:
create SRK(s) and the salting key for sessions using the
default well-known TPM password.
Some of these operations are lengthy, so performing them
earlier allows to minimize the remaining time for taking
the TPM ownership and improve UX.

BUG=chromium:772187
TEST=go through OOBE on eve

Change-Id: Id73b4252940c767fc2fa45414ed2ff7bdbec4887
Reviewed-on: https://chromium-review.googlesource.com/731857
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>