chromeos-config: Refactor cros_config to use a base offset
At present we use the full path to a node when reading properties from it.
This works correctly, but is inefficient since we repeat searches already
done at init time. It makes sense to conduct the search from the
/chromeos/models/MODEL node since we already have that available.

Refactor the code to handle this. There is no functional change, but this
permits us (in a later CL) to conduct the search relative to a different
base node.

BUG=chromium:777214
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools
Change-Id: Id09a9a6234a5afa94b105a6d5bed9485ffe94f88
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/732646