chromeos-config: Add a --abspath flag to cros_config
Use the target-directory map provided by the ebuild to make cros_config
aware of directory paths at run-time. This allows us to centralise the
use of paths in the config system rather than having the same path
repeated in the build system (e.g. ebuild) and the init scripts. We can
then be more sure that the correct path is used at run-time and thus that
the file will be found at run-time.

Update the Fake with this new interface.

BUG=chromium:769575
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Change-Id: I2c1693e6ee488ca5a16a1b981c0115aca6325d5c
Reviewed-on: https://chromium-review.googlesource.com/732645
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>