chromeos-config: Generate the target dirs config
At present the target directory information is in the schema and thus is
only accessible to the Python (host) side. It cannot be used at run-time.
We want to add support for cros_config to add the path to the values it
returns at run-time.

Add an option to cros_config_host to write this information to a .dtsi
file so that it can be included in the master configuration. In this way,
it can be accessed by cros_config.

BUG=chromium:769575
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: Ie53ec6a30ab772aff081bd77f516fff98b256f0d
Reviewed-on: https://chromium-review.googlesource.com/732321
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>