chromeos-config: Add a test for CrosConfigFake
We don't have a test for this at present which makes it vulnerable to
changes in the underlying interface and bugs in the implementation. At
present these might show up as failures in session_manager, for example.

Add a simple test for the Fake, representing our deep sympathy for our
clients.

BUG=chromium:774498
TEST=With a change to run this test from the ebuild:
FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Change-Id: Iceba0866e71a12578e1ae69a5f3769d562be15f2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/732643
Reviewed-by: Jason Clinton <jclinton@chromium.org>