chromeos-config: Fix validation errors in the test config
At present test.dts has various validation errors. Since it is only used
for unit testing, this is not really a problem. But it might be confusing
to someone who examines it closely. Also, running the validator on the
test file shows up these problems.

Fix these errors for consistency.

BUG=chromium:777214
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Change-Id: I19e22e59b9f8ef7fe58226d7782a75c90ea6106a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/732649
Reviewed-by: Jason Clinton <jclinton@chromium.org>