chromeos-config: Implement a 'default' feature
It is annoying to have to copy the configuration for one model into
another when they are almost identical. The existing sharing mechanisms
after sufficient for whitelabel and firmware, but do not cope well with
the case where two models are virtually the same.

Add a 'default' property, which allows all of the nodes and properties of
one node to default to those of another. This is an experimental feature.

So far the feature only exists in the host library. Further work is needed
to support it in the validator and the run-time library.

NOTE: This is probably taking things a little too far. We could end up
with  a situation where nothing is repeated in the config for a model. The
cost associated with this is that you must refer to other model to see the
actual settings. Once we have a feel for how useful/painful this is, we
can make a decisiion as to whether it stays.

BUG=chromium:777214
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Change-Id: I17243a59c3d6856b360cb6d630b103582bff1602
Reviewed-on: https://chromium-review.googlesource.com/732647
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>