chromeos-config: Implement submodel support in cros_config_host
Properties can be defined either in submodels or in the parent model. Add
support for reading a property from the submodel node if there is one.
Since on the host side we can read from any model or submodel, we need to
add submodel as a parameter.

BUG=chromium:777725
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Change-Id: Ie5e5bbcc0fa1b21b79fbc2c2ee502c72d1077f56
Reviewed-on: https://chromium-review.googlesource.com/740121
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>