chromeos-config: Correct URI list with shared firmware
This implementation is not actually correct at present. It assumes that
each model with shared firmware has it in its own location. In fact, the
images are truely shared, and exist in one place. Also we should not
return duplicates.

Fix these issues. Part of the problem comes from taking a model-centric
view of the URI list. In fact, the URI list is a global thing across all
models and it does not make much sense to request the URIs of a single
model. We never use that the firmware ebuild, and we are not allowed to
change the value of SRC_URI, so we have to have everything in the list
at all times.

Also tidy up the test data a little as it is confusing to have pyro
using an overlay named 'reef'.

BUG=chromium:775141
TEST=./run_tests.sh
FEATURES=test sudo -E  emerge chromeos-config-tools
PYTHONPATH=~/c python cros_config_host.py  -c \
	~/c/chroot/build/coral/usr/share/chromeos-config/config.dtb \
	 get-firmware-uris
See that we get the corect set:
gs://chromeos-binaries/HOME/bcs-coral-private/overlay-coral-private/\
	chromeos-base/chromeos-firmware-coral/Coral.9984.0.0.tbz2 \
gs://chromeos-binaries/HOME/bcs-coral-private/overlay-coral-private/\
	chromeos-base/chromeos-firmware-coral/Coral_EC.9984.0.0.tbz2
Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: Ib691d30953ad7e21146901d90f8e7ff1ef81b596
Reviewed-on: https://chromium-review.googlesource.com/736509
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>