imageloader: add GetComponentMetadata D-Bus method
This allows us to pack some information into the imageloader
manifest that will be signed but which we can access before we
actually load components. This information might be used to make
decisions about whether or not we actually do need to load the
component.

BUG=chromium:773471
TEST=deploy on device and test, new unit tests

Change-Id: Ia5e9f963e12dcd443851773f428d0d9f6c2104d8
Reviewed-on: https://chromium-review.googlesource.com/737428
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>