shill: cellular: simplify service storage identifier initialization
The storage identifier used by a CellularService object is supposed to
settle on a value before the CellularService object is registered with
the Manager. Currently, the storage identifier of a CellularService
object is initialized in the CellularService constructor and then set
again by the Cellular object before the CellularService object is
registered with the Manager. This CL moves all the storage identifier
initialization code to the CellularService constructor and removes
CellularService::SetStorageIdentifier.

BUG=None
TEST=Run unit tests.

Change-Id: I9681012eac5e8c73705a9dc828640b3d59390661
Reviewed-on: https://chromium-review.googlesource.com/737938
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>