chromeos-config: Allow selecting schema element types
When searching for a schema element the caller is often looking only for a
node, or only looking for a property. Allow the caller to specify this. It
allows us to catch a few more validation problems. Fix these in the README.

BUG=chromium:761265
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools
PYTHONPATH=~/c python validate_config.py README.md

Change-Id: Ic43a15fb4ca0dbc22529af98acbabbec0c8c969c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/739397
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>