chromeos-config: Add support for decoding SKU tables
At present mosys decodes the SKU table and provides us with a model name.
This is a useful feature since lots of things use mosys at present.
However we do want to move away from using mosys for identity at some
point.

On the other hand we don't necessarily want to read from SMBIOS tables and
VPD in cros_config. A reasonable compromise is to have mosys provide the
raw identity information (platform name and SKU ID) and cros_config to use
that to find the model. This has the advantage that we can support
determining the submodel and anything else we want to do with the
information.

To this end, implement SKU map decoding in cros_config. This is so far not
plumbed in. That (and tests) comes in a later CL.

BUG=chromium:777725,chromium:770768
CQ-DEPEND=CL:737802
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools
(which really just tests that it builds)

Change-Id: I3e95ed8a0906c06f398e3b4410f7993f19dc13ab
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/739398
Reviewed-by: Jason Clinton <jclinton@chromium.org>