chromeos-config: Switch to using the SKU for identification
Rather than CrosConfig calling 'mosys platform model' to find out the
current model, have it call 'mosys platform id' to get the raw information
from which it can figure it out itself. This allows it to decode the
submodel as well as the model.

Adjust the tests to use this same method. This requires the addition of a
SKU map into our test config.

BUG=chromium:777725
TEST=FEATURES=test emerge-reef-uni --nodeps chromeos-config-tools

Change-Id: Ib3563602492fe9a5f7336e4fbf67a170833d9281
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/739399
Reviewed-by: Jason Clinton <jclinton@chromium.org>