chromeos-config: Set up cras-config-dir as a target directory
We want to use this as a target directory in an init script. Mark it as
such in the schema. This allows 'cros_config --abspath' to work with it.

Also this has shown up a problem in the tests. The currently schema is
hard-coded in test.dts rather than being written from cros_config_host.
Update the test script to generate the file each time.

Finally, add a test for each file property that exists in the schema.

BUG=chromium:773898
TEST=FEATURES=test sudo -E emerge --nodeps chromeos-config-tools
manually on snappy

Change-Id: I5ea232aad9473f9086e479c1335a8f5caba96763
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/743025
Reviewed-by: Jason Clinton <jclinton@chromium.org>