chromeos-config: Refactor node init to handle whole tree
CrosConfig creates it own node /property tree parallel to the Fdt version.
This is so it can have its own methods and some abstraction from the
underlying Fdt.

At present the creation of the config tree is a bit unusual. It creates
nodes only for models, then locates the phandles separately in the
underlying Fdt. It has no knowledge of the family node, for example.

Adjust the init method to scan the entire tree. Phandles are located as
we progress and point to a CrosConfig Node. The same happens with
submodels.

Also drop the additional iteration in the CrosConfig constructor since we
can just use the root node. In fact it might be better for CrosConfig to
have a single root property instead of a nodes list, but that would
require a little more refactoring.

With these changes we have a consistent tree with all nodes present.

BUG=chromium:775229
TEST=./run_tests.sh

Change-Id: Ia7e319e35ad22aa09ca27256929ffa47d4be221f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/743028
Reviewed-by: Jason Clinton <jclinton@chromium.org>