chromeos-config: Move default master config into library
We should not have logic in the host tool for locating the master config
since users of the library cannot access it.

Move this code in to the library instead. Then both the host tool and any
library clients can obtain the master configuration without providing an
explicit config file.

BUG=chromium:775220
TEST=(in chroot) ./run_tests

Change-Id: Id0b4a933d1c7d9e99afe3261b1fc32fdcc3b0f2e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/742476