chromeos-config: Add support for BCS files
At present we require that all firmware files exist in the git repo. This
was a useful way of getting going, but is not desirable. Traditionally we
store larger files in BCS (Binary Cloud Storage) instead. At present
unibuild has no support for this.

Add support for reading touch firmware files from BCS instead of the local
filedir.

BUG=chromium:779438
TEST=FEATURES=test sudo -E emerge --nodeps chromeos-config-host

Change-Id: I6c7e409cfdcd4a156c91d0df3d2136f2d1466790
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/742305
Reviewed-by: Jason Clinton <jclinton@chromium.org>