libpasswordprovider: Initial checkin for password provider library.
This is the initial checkin for a simple static library that stores
the user's password in a kernel keyring. This is the first checkin
for a feature that proxies the user's password to the wifi config
to be used for authentication.

Detailed design doc at: go/cros-password-provider-dd

BUG=chromium:386606
TEST=run unit tests

Change-Id: I61f03413b0dcf747bda309bc254b8cfc6cd50852
Reviewed-on: https://chromium-review.googlesource.com/748024
Commit-Ready: May Lippert <maybelle@chromium.org>
Tested-by: May Lippert <maybelle@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>