shill: rename Profile::set_storage() to Profile::SetStorageForTest()
Profile::set_storage() is provided for the purpose of unit testing. This
CL renames it to Profile::SetStorageForTest() to better indicate that.
It also changes the method to take the StorageInterface object by
std::unique_ptr instead of raw pointer to signify the ownerhsip
transfer.

BUG=None
TEST=Run unit tests.

Change-Id: I185f3cfe7a8bc2230be54e992fb5f6324c2d1854
Reviewed-on: https://chromium-review.googlesource.com/748269
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>