chromeos-config: audio: Support submodels properly
At present the submodel config is not actually scanned when generating the
audio file list. This was missed when adding the submodel implementation.

Scan all models / submodels and generate audio config for each. At present
the config must either all be in the model or all be in the submodels.
However, since the validator checks that, we can just pick up everything
we can find.

BUG=chromium:773898
TEST=./run_tests.sh
(with updated tests)

Change-Id: If34546f8af84562c6b7b32f4b0267f1c94672084
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/753523
Reviewed-by: Jason Clinton <jclinton@chromium.org>