chromeos-config: audio: Allow more properties in the family
In some cases we share these across models and it is convenient to put
cras-config-dir, ucm-suffix and topology-bin in the family instead of the
model.

This was shown up recently in the coral config where submodels are used to
provide 1- or 2-microphone configuration, and it is easier to specify the
details once in the family.

Putting this config in the family is not mandatory, but these properties
must be present in either the model node or the family. At present there
is no validation to check this. Assuming we do continue to want to support
both, we should add this.

Also update the test to have this more realistic example.

BUG=chromium:773898
TEST=./run_tests.sh
(with updated tests)

Change-Id: I0a6ee6fe61dd1e675fb0db925714389d15d25a4d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/753524
Reviewed-by: Jason Clinton <jclinton@chromium.org>