userfeedback: simplify info collection
The `echo -n` usage here is not portable, and it's pointless as it
doesn't emit any additional lines.

Create a for loop to avoid copying & pasting all the same things,
and prefix each line with the corresponding field so it's clear what
each line actually means.

BUG=None
TEST=mosys_info looks nicer

Change-Id: If0076e3c742b1969e542b2cc1400211979796606
Reviewed-on: https://chromium-review.googlesource.com/753749
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>