intel: ipu3: Use CLOCK_MONOTONIC instead CLOCK_REALTIME in 3A Client
1. When the system sleeps, the system time will not stop.
So if the sleep happens,
the pthread_cond_timedwait() may timeout.
So we use the CLOCK_MONOTONIC instead of CLOCK_REALTIME.
2. Make the timeout value to be 5 seconds.
From the obversation for the s0/s3 testing,
when the sleep happens and it happens to the waiting
callback is called,
the callback could be more than 2 seconds sometimes,
even the remote side has returned very quickly.

BUG=b:62358788
TEST=Basic camera functions work on soraka.

Change-Id: I51dd293e9eed185de121eb9cd6a6237853e56c2d
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/758299
Commit-Ready: WeiX Meng <weix.meng@intel.com>
Tested-by: WeiX Meng <weix.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>