libcros_config: Add access to firmware information
We want to adjust pack_firmware so that it does not directly read the
master configuration itself, but rather go through CrosConfig.

Add a function which figures out the information needed by the firmware
packer and returns it. Some of the config has to be processed to produce
the required fields.

This should give us a good basis for adjusting the schema as needed. For
example we want to support an alternative whitelabel schema.

BUG=chromium:775220
TEST=PYTHONPATH=~/c python libcros_config_host/fdt_unittest.py

Change-Id: Id201af1a1949ae4f8ef6ac7c408638ddb02be311
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/760441
Commit-Ready: Lann Martin <lannm@chromium.org>
Tested-by: Lann Martin <lannm@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Lann Martin <lannm@chromium.org>