cryptohome: delete can_attempt_ownership after initialization
Before this change, the persistent can_attempt_ownership flag
was set by cryptohomed after receiving CanAttemptOwnership from
Chrome. The flag was deleted only during powerwash. The flag is
required to re-attempt initializing the TPM if the initialization
was interrupted.

This behavior required developers who don't perform the powerwash
when clearing the owner during developing or debugging to delete
the persistent flag file. Otherwise, the TPM was re-initialized
after clearing the owner without waiting for OOBE signal.

This change deletes the persistent flag after the TPM is
successfully initialized (and thus we don't need it anymore to
re-attempt initialization). This prevents automatic re-initialization
of the TPM in the developers' workflow, if initialization was
completed (or didn't start) before clearing the owner. This is
the typical case, and the majority of the developers won't need
to delete the flag file anymore.

To achieve that:
1) The persistent flag logic was moved from Service to a (recently
   created) TpmPersistentState, which is a better place for it in
   any case. While moving, the flag was renamed from
   CanAttemptOwnership to ShallInitialize, what better represents
   its meaning and usage in cryptohomed.
2) An ability to clear the flag was added to the interface.
3) TpmInit now both sets the flag when it attempts a TPM
   initialization and clears the flag after successful initialization.

BUG=none
TEST=unit tests

Change-Id: I53ead2b7ac2654f1b7d1893c86a5164c384a243b
Reviewed-on: https://chromium-review.googlesource.com/759697
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>