installer: setgoodkernel rolls forward as far as possible
On the boot following chromeos-setgoodkernel marking a kernel as good,
RW firmware will roll forward the minimum allowable kernel vboot
version stored in the TPM as far as possible.

This is the correct behavior on unenrolled devices, and duplicates the
behavior of older firmware which does not support
kernel_max_rollforward.

Note that the firmware change which uses this setting is dependent on
this change, to make sure current behavior is maintained when that
change commits.

Once enterprise support for kernel_max_rollforward policy has been
added, this script should be changed so it only forces max rollforward
on unenrolled devices.  Added a TODO to that effect.

BUG=chromium:783997
BRANCH=none
CQ-DEPEND=CL:765572
TEST=manual
	# Overrides value of 0
	crossystem kernel_max_rollforward=0
	chromeos-setgoodkernel
	crossystem kernel_max_rollforward --> 0xfffffffe

	# Overrides non-zero values too
	crossystem kernel_max_rollforward=0x10001
	chromeos-setgoodkernel
	crossystem kernel_max_rollforward --> 0xfffffffe

Change-Id: I8e3488553569ade3fea0c885e18b6daf0b5663aa
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/765573
Reviewed-by: Stefan Reinauer <reinauer@google.com>