vm_tools: Mount nfs inside VM
Automatically mount the shared NFS directory inside the VM if one is
provided to vm_launcher.  Also, since nfs-ganesha's lifetime is tied to
that of rpcbind, change the Start and Stop functions to start and stop
rpcbind instead.  The upstart config for nfs-ganesha will ensure that it
automatically gets started/stopped as necessary.

BUG=none
TEST=vm_launcher run --rwcontainer=/mnt/stateful_partition/stretch.img \
                     --nfs=/home/chronos/user/ \
                     --vm_path=/mnt/stateful_partition/vm/ \
                     debian
CQ-DEPEND=CL:762129

Change-Id: I60994114c7069d12cf1bd5b50495f421818ad837
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/768109